+++ src/MediaHandler/FileHandler.php
@@ -9,10 +9,10 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Outl1ne\NovaMediaHub\MediaHandler\Support\Base64File;
 use Outl1ne\NovaMediaHub\MediaHandler\Support\RemoteFile;
 use Outl1ne\NovaMediaHub\MediaHandler\Support\Filesystem;
-use Outl1ne\NovaMediaHub\Exceptions\FileTooLargeException;
 use Outl1ne\NovaMediaHub\MediaHandler\Support\FileHelpers;
 use Outl1ne\NovaMediaHub\Jobs\MediaHubOptimizeAndConvertJob;
 use Outl1ne\NovaMediaHub\Exceptions\NoFileProvidedException;
+use Outl1ne\NovaMediaHub\Exceptions\FileValidationException;
 use Outl1ne\NovaMediaHub\Exceptions\UnknownFileTypeException;
 use Symfony\Component\HttpFoundation\File\File as SymfonyFile;
 use Outl1ne\NovaMediaHub\Exceptions\FileDoesNotExistException;
@@ -32,11 +32,11 @@ class FileHandler
     protected string $conversionsDiskName = '';
     protected string $collectionName = '';
     protected array $modelData = [];
+    protected bool $deleteOriginal = false;
 
     public function __construct()
     {
         $this->filesystem = app()->make(Filesystem::class);
-        $this->fileNamer = config('nova-media-hub.file_namer');
     }
 
     public static function fromFile($file): self
@@ -62,9 +62,13 @@ class FileHandler
         }
 
         if ($file instanceof UploadedFile) {
-            $this->pathToFile = $file->getPath() . '/' . $file->getFilename();
-            $this->fileName = $file->getClientOriginalName();
-            return $this;
+            if ($file->getError()) {
+                throw new FileValidationException($file->getErrorMessage());
+            } else {
+                $this->pathToFile = $file->getPath() . '/' . $file->getFilename();
+                $this->fileName = $file->getClientOriginalName();
+                return $this;
+            }
         }
 
         if ($file instanceof SymfonyFile) {
@@ -82,6 +86,12 @@ class FileHandler
 
         $this->file = null;
         throw new UnknownFileTypeException();
+    }
+
+    public function deleteOriginal($deleteOriginal = true)
+    {
+        $this->deleteOriginal = $deleteOriginal;
+        return $this;
     }
 
     public function withCollection(string $collectionName)
@@ -108,12 +118,24 @@ class FileHandler
         return $this;
     }
 
-    public function save($file = null): Media
+    public function save($file = null): ?Media
     {
         if (!empty($file)) $this->withFile($file);
         if (empty($this->file)) throw new NoFileProvidedException();
         if (!is_file($this->pathToFile)) throw new FileDoesNotExistException($this->pathToFile);
 
+        try {
+            return $this->saveFile();
+        } finally {
+            // Ensure cleanup
+            if ($this->deleteOriginal && is_file($this->pathToFile)) {
+                unlink($this->pathToFile);
+            }
+        }
+    }
+
+    private function saveFile(): ?Media
+    {
         // Check if file already exists
         $fileHash = FileHelpers::getFileHash($this->pathToFile);
         $existingMedia = MediaHub::getQuery()->where('original_file_hash', $fileHash)->first();
@@ -121,28 +143,34 @@ class FileHandler
             $existingMedia->updated_at = now();
             $existingMedia->save();
             $existingMedia->wasExisting = true;
-            return $existingMedia;
-        }
 
-        // Check file size
-        $maxSizeBytes = MediaHub::getMaxFileSizeInBytes();
-        if ($maxSizeBytes && filesize($this->pathToFile) > $maxSizeBytes) {
-            throw new FileTooLargeException($this->pathToFile);
+            // Delete original
+            if ($this->deleteOriginal && is_file($this->pathToFile)) {
+                unlink($this->pathToFile);
+            }
+
+            return $existingMedia;
         }
 
         $sanitizedFileName = FileHelpers::sanitizeFileName($this->fileName);
         [$fileName, $rawExtension] = FileHelpers::splitNameAndExtension($sanitizedFileName);
         $extension = File::guessExtension($this->pathToFile) ?? $rawExtension;
+        $mimeType = File::mimeType($this->pathToFile);
+        $fileSize = File::size($this->pathToFile);
 
         $this->fileName = MediaHub::getFileNamer()->formatFileName($fileName, $extension);
+
+        // Validate file
+        $fileValidator = MediaHub::getFileValidator();
+        $fileValidator->validateFile($this->collectionName, $this->pathToFile, $this->fileName, $extension, $mimeType, $fileSize);
 
         $mediaClass = MediaHub::getMediaModel();
         $media = new $mediaClass($this->modelData ?? []);
 
         $media->file_name = $this->fileName;
         $media->collection_name = $this->collectionName;
-        $media->size = File::size($this->pathToFile);
-        $media->mime_type = File::mimeType($this->pathToFile);
+        $media->size = $fileSize;
+        $media->mime_type = $mimeType;
         $media->original_file_hash = $fileHash;
         $media->data = [];
         $media->conversions = [];
@@ -155,7 +183,7 @@ class FileHandler
 
         $media->save();
 
-        $this->filesystem->copyFileToMediaLibrary($this->pathToFile, $media, $this->fileName, Filesystem::TYPE_ORIGINAL, true);
+        $this->filesystem->copyFileToMediaLibrary($this->pathToFile, $media, $this->fileName, Filesystem::TYPE_ORIGINAL, $this->deleteOriginal);
 
         MediaHubOptimizeAndConvertJob::dispatch($media);