+++ src/Models/Media.php
@@ -5,7 +5,6 @@ namespace Outl1ne\NovaMediaHub\Models;
 use Outl1ne\NovaMediaHub\MediaHub;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Storage;
-use Outl1ne\NovaMediaHub\MediaHandler\Support\FileNamer;
 
 class Media extends Model
 {
@@ -39,18 +38,22 @@ class Media extends Model
 
     public function getUrlAttribute()
     {
-        return Storage::disk($this->disk)->url($this->path . FileNamer::encode($this->file_name));
+        return Storage::disk($this->disk)->url($this->path . $this->file_name);
+    }
+
+    public function getUrl(?string $forConversion = null)
+    {
+        if (!$forConversion) return $this->url;
+
+        $conversionName = $this->conversions[$forConversion] ?? null;
+        if (empty($conversionName)) return null;
+
+        return Storage::disk($this->conversions_disk)->url($this->conversionsPath . $conversionName);
     }
 
     public function getThumbnailUrlAttribute()
     {
-        $thumbnailConversionName = MediaHub::getThumbnailConversionName();
-        if (!$thumbnailConversionName) return null;
-
-        $thumbnailName = $this->conversions[$thumbnailConversionName] ?? null;
-        if (!$thumbnailName) return null;
-
-        return Storage::disk($this->conversions_disk)->url($this->conversionsPath . FileNamer::encode($thumbnailName));
+        return $this->getUrl(MediaHub::getThumbnailConversionName());
     }
 
     public function formatForNova()